requirements.txt
python-telegram-bot>=22.0
SQLAlchemy>=2.0.23
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
pillow>=10.1.0 
Flask>=3.0.0
requests>=2.31.0
pyngrok>=7.0.0 
matplotlib>=3.8.0
seaborn>=0.13.0
waitress>=3.0.0
psutil>=5.9.0 